worker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
pika
redis
transformers
torch
psycopg2-binary